requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
crewai>=0.28.0
crewai-tools>=0.1.0
tavily-python>=0.3.0
pydantic>=2.0.0
openai>=1.0.0